Revert "refactored palmpay service class"
