+++ app/Helpers/GeneralHelpers.php
@@ -86,9 +86,7 @@ class GeneralHelpers{
         }
     }
 
-    public static function randomDelay($seconds = 10) : void
-    {
-        $delay = rand(1, $seconds);
-        sleep($delay);
-    }
+
+    
+
 }
