+++ app/Http/Controllers/V1/PalmPayWebhookController.php
@@ -4,12 +4,12 @@ namespace App\Http\Controllers\V1;
 
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
-use App\Services\Money\PalmPayFundingService;
+use App\Services\Money\PalmPayService;
 
 class PalmPayWebhookController extends Controller
 {
     public function __invoke(Request $request)
     {
-        return PalmPayFundingService::webhook($request);
+        return PalmPayService::webhook($request);
     }
 }
